Detect breaks using an integer h parameter
Due to varying numbers of NA values, we don't get a consistent
minimum segment size, because 11% of 2 observations is not the same
as 11% of 200 observations. This led to too short minimum segment
sizes and thus multiple breaks being detected during a year.

NB: sctest can't handle integers there, so for the test we still
use the old relative method.